tokenizer functions reassemble
