+++ tokenizer/token_checker.c
@@ -1,0 +1,39 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   token_checker.c                                    :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: rbarkhud <rbarkhud@student.42yerevan.am    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/11/05 22:21:01 by rbarkhud          #+#    #+#             */
+/*   Updated: 2025/11/05 22:21:44 by rbarkhud         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "./tokenizer.h"
+
+int	ft_is_word_part(char c)
+{
+	return (ft_isalnum(c) || ft_inset(c, "!@#$%^*-_+=~`;.?,{}[]\\/"));
+}
+
+int	ft_isspace(char c)
+{
+	return ((c >= 9 && c <= 13) || c == ' ');
+}
+
+int	ft_inset(char c, char *set)
+{
+	int	i;
+
+	if (!set)
+		return (0);
+	i = 0;
+	while (set[i])
+	{
+		if (c == set[i])
+			return (1);
+		++i;
+	}
+	return (0);
+}
