+++ tokenizer/tokenizer_utils.c
@@ -6,7 +6,7 @@
 /*   By: rbarkhud <rbarkhud@student.42yerevan.am    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/16 20:10:31 by rbarkhud          #+#    #+#             */
-/*   Updated: 2025/11/05 21:35:16 by rbarkhud         ###   ########.fr       */
+/*   Updated: 2025/11/05 22:20:57 by rbarkhud         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -73,25 +73,4 @@ int	get_token_type(char *value)
 	else if (parenthesis_type != -1 && ft_inset(*value, "<>()"))
 		return (parenthesis_type);
 	return (TK_WORD);
-}
-
-int	ft_isspace(char c)
-{
-	return ((c >= 9 && c <= 13) || c == ' ');
-}
-
-int	ft_inset(char c, char *set)
-{
-	int	i;
-
-	if (!set)
-		return (0);
-	i = 0;
-	while (set[i])
-	{
-		if (c == set[i])
-			return (1);
-		++i;
-	}
-	return (0);
 }
